requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==5.0.0
billiard==4.2.1
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.2
django-model-utils==5.0.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
flake8==7.1.1
gunicorn==23.0.0
inflection==0.5.1
kombu==5.4.2
mccabe==0.7.0
packaging==24.1
pillow==11.0.0
prompt_toolkit==3.0.48
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
